refactor(db): move reading database functions to db directory
- Move getReading and getReadings from app/readings/utils.ts to lib/db/readings.ts
- Update imports in all affected files
- Keep utils.ts temporarily with re-exports for backward compatibility
- Update lib/db/index.ts to export the functions
- Add JSDoc comments to functions
- Update TODO.md to mark task as completed
