+++ src/app/readings/utils.ts
@@ -1,50 +1,11 @@
-import prisma from '@/lib/prisma'
+/**
+ * @deprecated This file is deprecated and will be removed in a future version.
+ * Please import from '@/lib/db' instead.
+ * 
+ * Example:
+ * import { getReading, getReadings } from '@/lib/db'
+ */
 
-export async function getReading(slug: string) {
-  try {
-    console.log(`Fetching reading with slug: ${slug}`)
-    
-    // Use raw query for maximum compatibility
-    const readings = await prisma.$queryRaw`
-      SELECT id, slug, title, author, "finishedDate", "coverImageSrc", thoughts, dropped
-      FROM "Reading"
-      WHERE slug = ${slug}
-      LIMIT 1;
-    `
-    
-    const reading = Array.isArray(readings) && readings.length > 0 ? readings[0] : null
-    
-    console.log(reading ? `Found reading: ${reading.title}` : `No reading found for slug: ${slug}`)
-    return reading
-  } catch (error) {
-    console.error(`Error fetching reading with slug ${slug}:`, error)
-    return null
-  }
-}
+import { getReading, getReadings } from '@/lib/db'
 
-export async function getReadings() {
-  try {
-    console.log('Getting readings from database...')
-    
-    // Use raw query for maximum compatibility
-    const readings = await prisma.$queryRaw`
-      SELECT id, slug, title, author, "finishedDate", "coverImageSrc", thoughts, dropped
-      FROM "Reading"
-      ORDER BY 
-        CASE WHEN "finishedDate" IS NULL THEN 1 ELSE 0 END,
-        "finishedDate" DESC,
-        id DESC;
-    `
-    
-    console.log(`Found ${Array.isArray(readings) ? readings.length : 0} readings`)
-    
-    if (!readings || (Array.isArray(readings) && readings.length === 0)) {
-      console.warn('No readings found in database')
-    }
-    
-    return readings
-  } catch (error) {
-    console.error('Error fetching readings:', error)
-    return []
-  }
-}
+export { getReading, getReadings }
